move ServerOwner to server/player
