+++ prc/models/server/status.py
@@ -1,34 +1,10 @@
 from typing import TYPE_CHECKING
 from enum import Enum
+from .player import ServerOwner
 
 if TYPE_CHECKING:
     from prc.server import Server
     from prc.api_types.v1 import ServerStatusResponse
-
-
-class ServerOwner:
-    """Represents a server [co-]owner partial player."""
-
-    def __init__(self, server: "Server", id: int):
-        self._server = server
-
-        self.id = int(id)
-
-    @property
-    def player(self):
-        """The full server player, if found."""
-        return self._server._get_player(id=self.id)
-
-    def __eq__(self, other: object) -> bool:
-        if isinstance(other, ServerOwner):
-            return self.id == other.id
-        return False
-
-    def __ne__(self, other: object) -> bool:
-        return not self.__eq__(other)
-
-    def __repr__(self) -> str:
-        return f"<{self.__class__.__name__} id={self.id}>"
 
 
 class AccountRequirement(Enum):
